Isolate get webview function to reuse
